feat: abstract urgency calculations into a separate utils file, import relevant functions into hooks.js, add relevant imports to index.js
