+++ src/utils/index.js
@@ -1,4 +1,5 @@
 export * from './dates';
 export * from './hooks';
 export * from './normalize';
+export * from './calculateUrgency';
 export * from './calculateIsPurchased';
